requirements.txt
Flask
python-dotenv
psycopg2-binary
sqlalchemy
pydantic
qrcode[pil]
requests
pillow
